requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
greenlet==2.0.2
idna==3.4
mysql-connector==2.2.9
numpy==1.26.0
pandas==2.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
